% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abpwaveforms-package.R
\docType{package}
\name{abpwaveforms-package}
\alias{abpwaveforms}
\alias{abpwaveforms-package}
\title{abpwaveforms: Arterial Blood Pressure Waveform Feature Extraction}
\description{
Rcpp-backed tools for ABP/ECG/PPG fiducials and features
(pulse pressure, AIx, RWTT, PAT, HR, etc.).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
