% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{process_condition}
\alias{process_condition}
\title{Process a CSV with ABP/ECG/PPG columns}
\usage{
process_condition(file, fs = 200)
}
\arguments{
\item{file}{CSV path containing columns ABP, ECG, PPG}

\item{fs}{Target Hz (default 200)}
}
\value{
Data frame of features
}
\description{
Process a CSV with ABP/ECG/PPG columns
}
